Task2

+++ main.h
@@ -2,6 +2,6 @@
 #define PRINTF_H
 int _printf(const char *format, ...);
 int _func(const char *ptr, int l, int *ctr);
-int get_length(int num);
-int intToString(int, char*);
+int get_length(long int num);
+int intToString(long int, char*);
 #endif
